Move insert_xai into separate functional api module
