+++ openvino_xai/__init__.py
@@ -7,6 +7,6 @@ OpenVINO-XAI library for explaining OpenVINO™ IR models.
 
 from .common.parameters import Method, Task
 from .explainer.explainer import Explainer
-from .inserter import insert_xai
+from .api.api import insert_xai
 
 __all__ = ["Explainer", "insert_xai", "Method", "Task"]
