+++ openvino_xai/api/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (C) 2023-2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+"""
+Finctional API.
+"""
+from openvino_xai.api.api import insert_xai
+
+__all__ = [
+    "insert_xai",
+]
