+++ tests/unit/insertion/test_insertion.py
@@ -4,7 +4,7 @@
 import pytest
 from openvino import runtime as ov
 
-from openvino_xai import insert_xai
+from openvino_xai.api.api import insert_xai
 from openvino_xai.common.parameters import Method, Task
 from openvino_xai.common.utils import has_xai, retrieve_otx_model
 from openvino_xai.inserter.parameters import DetectionInsertionParameters
